# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: FinBridge API - Open Banking + Financial Inclusion platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/me:
    get:
      operationId: getMe
      tags: [users]
      summary: Get current user profile
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    patch:
      operationId: updateMe
      tags: [users]
      summary: Update user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/me/kyc:
    post:
      operationId: submitKyc
      tags: [users]
      summary: Submit KYC upgrade request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KycInput"
      responses:
        "200":
          description: KYC result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  # ── ACCOUNTS ──────────────────────────────────────────────────────────────
...
  /transactions:
    get:
      operationId: listTransactions
      tags: [transactions]
      summary: List all transactions (paginated)
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
        - name: accountId
          in: query
          schema:
            type: ["integer", "null"]
        - name: category
          in: query
          schema:
            type: ["string", "null"]
        - name: type
          in: query
          schema:
            type: ["string", "null"]
            enum: [credit, debit, null]
      responses:
        "200":
          description: Paginated transactions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionList"
    post:
      operationId: createTransaction
      tags: [transactions]
      summary: Record a manual transaction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransactionInput"
      responses:
        "201":
          description: Created transaction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
...
  /savings:
    get:
      operationId: listSavingsGoals
      tags: [savings]
      summary: List all savings goals
      responses:
        "200":
          description: Savings goals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SavingsGoal"
    post:
      operationId: createSavingsGoal
      tags: [savings]
      summary: Create a new savings goal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SavingsGoalInput"
      responses:
        "201":
          description: Created savings goal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SavingsGoal"
  /savings/{id}:
    get:
      operationId: getSavingsGoal
      tags: [savings]
      summary: Get a single savings goal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Savings goal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SavingsGoal"
    patch:
      operationId: updateSavingsGoal
      tags: [savings]
      summary: Update a savings goal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SavingsGoalUpdate"
      responses:
        "200":
          description: Updated savings goal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SavingsGoal"
    delete:
      operationId: deleteSavingsGoal
      tags: [savings]
      summary: Delete a savings goal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessMessage"
  /savings/{id}/contribute:
    post:
      operationId: contributeToGoal
      tags: [savings]
      summary: Add money to a savings goal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContributionInput"
      responses:
        "200":
          description: Updated savings goal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SavingsGoal"

  # ── COMMUNITY GROUPS (AJO/ESUSU) ──────────────────────────────────────────
...
components:
  schemas:
...
    SuccessMessage:
      type: object
      properties:
        message:
          type: string
      required:
        - message

    # ── USER ──────────────────────────────────────────────────────────────
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        bvn:
          type: ["string", "null"]
        kycTier:
          type: string
          enum: [basic, intermediate, full]
        financialHealthScore:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - email
        - kycTier
        - financialHealthScore
        - createdAt
    UserUpdate:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
    KycInput:
      type: object
      properties:
        tier:
          type: string
          enum: [intermediate, full]
        bvn:
          type: ["string", "null"]
        nin:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
      required:
        - tier

    # ── ACCOUNTS ──────────────────────────────────────────────────────────
...
    Transaction:
      type: object
      properties:
        id:
          type: integer
        accountId:
          type: integer
        amount:
          type: number
        type:
          type: string
          enum: [credit, debit]
        category:
          type: string
        description:
          type: string
        date:
          type: string
          format: date-time
        status:
          type: string
          enum: [completed, pending, failed]
        bankName:
          type: ["string", "null"]
      required:
        - id
        - accountId
        - amount
        - type
        - category
        - description
        - date
        - status
    TransactionInput:
      type: object
      properties:
        accountId:
          type: integer
        amount:
          type: number
        type:
          type: string
          enum: [credit, debit]
        category:
          type: string
        description:
          type: string
        date:
          type: string
          format: date-time
      required:
        - accountId
        - amount
        - type
        - category
        - description
    TransactionList:
      type: object
      properties:
        transactions:
          type: array
          items:
            $ref: "#/components/schemas/Transaction"
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
      required:
        - transactions
        - total
        - limit
        - offset
...
    SavingsGoal:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        targetAmount:
          type: number
        currentAmount:
          type: number
        deadline:
          type: ["string", "null"]
          format: date
        autoDeductAmount:
          type: ["number", "null"]
        autoDeductEnabled:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - targetAmount
        - currentAmount
        - autoDeductEnabled
        - createdAt
    SavingsGoalInput:
      type: object
      properties:
        name:
          type: string
        targetAmount:
          type: number
        deadline:
          type: ["string", "null"]
          format: date
        autoDeductAmount:
          type: ["number", "null"]
        autoDeductEnabled:
          type: boolean
      required:
        - name
        - targetAmount
    SavingsGoalUpdate:
      type: object
      properties:
        name:
          type: string
        targetAmount:
          type: number
        deadline:
          type: ["string", "null"]
          format: date
        autoDeductAmount:
          type: ["number", "null"]
        autoDeductEnabled:
          type: boolean
    ContributionInput:
      type: object
      properties:
        amount:
          type: number
      required:
        - amount

    # ── COMMUNITY GROUPS ──────────────────────────────────────────────────
...