# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * FinBridge API - Open Banking + Financial Inclusion platform
 * OpenAPI spec version: 0.1.0
 */
import type { UserKycTier } from './userKycTier';

export interface User {
  id: number;
  name: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  bvn?: string | null;
  kycTier: UserKycTier;
  financialHealthScore: number;
  createdAt: Date;
}