# lib/api-zod/src/generated/types/userUpdate.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * FinBridge API - Open Banking + Financial Inclusion platform
 * OpenAPI spec version: 0.1.0
 */

export interface UserUpdate {
  name?: string;
  phone?: string;
}